Rename status endpoint and update references
